+++ packages/core/lib/src/core_widget_factory.dart
@@ -11,6 +11,7 @@ import 'internal/core_ops.dart';
 import 'internal/core_parser.dart';
 import 'internal/flattener.dart';
 import 'internal/margin_vertical.dart';
+import 'internal/ops/flex.dart';
 import 'internal/platform_specific/fallback.dart'
     if (dart.library.io) 'internal/platform_specific/io.dart';
 
@@ -686,7 +687,7 @@ class WidgetFactory {
         final displayExpression = displayExpressions.isNotEmpty ? displayExpressions.first : null;
         final display = displayExpression?.term ?? 'block';
         if (display == 'flex') {
-          meta.register(_flexOp(meta));
+          meta.register(FlexOps.flexOp(meta));
         }
 
         meta[kCssDisplay] = kCssDisplayBlock;
@@ -1156,99 +1157,6 @@ class WidgetFactory {
     }
 
     return baseUrl.resolveUri(uri).toString();
-  }
-
-  /// Builds custom widget for div elements with display: flex from [meta]
-  BuildOp _flexOp(BuildMetadata meta) {
-    return BuildOp(
-      onChild: (childMeta) {
-        childMeta.register(_flexItemOp(childMeta));
-      },
-      onWidgets: (meta, widgets) {
-        final String id = meta.element.id;
-        String flexDirection = 'row';
-        String justifyContent = 'flex-start';
-        String alignItems = 'flex-start';
-
-        for (final element in meta.element.styles) {
-          final String? value = element.term;
-
-          if (value != null) {
-            switch (element.property) {
-              case 'flex-direction':
-                flexDirection = value;
-              break;
-              case 'justify-content':
-                justifyContent = value;
-              break;
-              case 'align-items':
-                alignItems = value;
-              break;
-            }
-          }
-        }
-
-        return [
-          Flex(
-            key: Key(id),
-            direction: 'row' == flexDirection ? Axis.horizontal : Axis.vertical,
-            mainAxisAlignment: _toMainAxisAlignment(justifyContent),
-            crossAxisAlignment: _toCrossAxisAlignment(alignItems),
-            children: widgets.toList()
-          )
-        ];
-      },      
-    );
-  }
-
-  /// Build op for child elements of flex containers
-  BuildOp _flexItemOp(BuildMetadata meta) {
-    return BuildOp(
-      defaultStyles: (element) {
-        return {
-          kCssWidth: "auto",
-          kCssHeight: "auto"
-        };
-      }
-    );
-  }
-
-  /// Converts CSS [justifyContent] to Flutter Grid MainAxisAlignment
-  static MainAxisAlignment _toMainAxisAlignment(String justifyContent) {
-    switch (justifyContent) {
-      case 'flex-start':
-        return MainAxisAlignment.start;
-      case 'flex-end':
-        return MainAxisAlignment.end;
-      case 'center':
-        return MainAxisAlignment.center;
-      case 'space-between':
-        return MainAxisAlignment.spaceBetween;
-      case 'space-around':
-        return MainAxisAlignment.spaceAround;
-      case 'space-evenly':
-        return MainAxisAlignment.spaceEvenly;
-      default:
-        return MainAxisAlignment.start;
-    }
-  }
-
-  /// Converts CSS [alignItems] to Flutter Grid CrossAxisAlignment
-  static CrossAxisAlignment _toCrossAxisAlignment(String alignItems) {
-    switch (alignItems) {
-      case 'flex-start':
-        return CrossAxisAlignment.start;
-      case 'flex-end':
-        return CrossAxisAlignment.end;
-      case 'center':
-        return CrossAxisAlignment.center;
-      case 'baseline':
-        return CrossAxisAlignment.baseline;
-      case 'stretch':
-        return CrossAxisAlignment.stretch;
-      default:
-        return CrossAxisAlignment.start;
-    }
   }
 
   BuildOp _anchorOp(String id) {
